resolve conflicts and move to dedicated  file
